% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_draws.R
\name{extract_draws}
\alias{extract_draws}
\title{Function for extracting Bayesian B-spline posteriors}
\usage{
extract_draws(
  fit,
  df,
  pars = c("adj", "sigma_ref", "sigma_obs", "tau", "C14_pred")
)
}
\arguments{
\item{fit}{cmdstan fit object of a Bayesian B-spline model}

\item{df}{data prepared using the data_prep function used to fit the model}

\item{pars}{list of parameters to extract. defaults to c('adj','sigma_ref','sigma_obs','tau','C14_pred')}
}
\value{
a list containing draws_matrix of posterior draws for parameters of interest; parameters return depend on the model fit
}
\description{
wrapper function for extracting posterior samples from the Bayesian B-spline model
}
\examples{
#reference-only
df <- data_prep(sim_ref)

#estimate model
fit <- est_model(df, show_messages = FALSE, show_exceptions = FALSE)

#extract draws
draws <- extract_draws(fit)

}
