% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plotting  Bayesian B-spline model}
\usage{
plot_fit(
  df,
  ext,
  probs = c(0.05, 0.5, 0.95),
  post.den = TRUE,
  legend = TRUE,
  min.BY,
  max.BY,
  inset.control = list(width = 0.003, height = 0.125, sig_low = 0.3, adj_low = 0.05,
    xbump = 0, lab_adj = 0.18)
)
}
\arguments{
\item{df}{data list object from \code{data_prep} and used to estimate the STAN model}

\item{ext}{posterior draws extraction list from \code{extract_draws} function}

\item{probs}{vector of three probabilities for lower CI, centrality, upper CI}

\item{post.den}{logical flag to plot posterior densities of certain parameters}

\item{legend}{logical flag to plot legend}

\item{inset.control}{a list with values of width (width of the inset posterior in proportions of xlim), height (height of the inset posterior in proportions of ylim), sig_low (lower position of sigma posterior in proportions of ylim), adj_low (lower position of BY_adj posterior in proportions of ylim), xbump (translation of posteriors along x-axis in proportions of xlim).}
}
\value{
figure only, invisible
}
\description{
function for plotting  Bayesian B-spline model for estimating reference series and/or assessing aging bias
}
\examples{
#REFERENCE-ONLY
df <- data_prep(sim_ref)

#estimate model
fit <- est_model(df, show_messages = FALSE, show_exceptions = FALSE)

#extract draws
draws <- extract_draws(fit)

#plot fit
plot_fit(df, draws)

#INTEGRATED
\dontrun{
df_int <- data_prep(sim_ref, sim_unk)
fit_int <- est_model(df_int)
draws_int <- extract_draws(fit_int)
plot_fit(df_int, draws_int)
}

}
